Inheritance
I have implemented an emitter class and a surfaceEmitter class and changes all the surface reports eg fgreport csreport btreports to use it.

So now the reports are really simple and provided taht we follow the same patter for all of them we can add them very easilly.

I have kept them as seperate modules for now rather than creating a generic process because I would like to be able to work in some UI elements.

But in theory we could create some automation that would allow us to populate new models without code changes