New API
The emitter now fetches the url of the server bu checking the live server for the client route.

If we are beta then it uses staging. If client route endoint doesnt exist of the client version doesnt have a route then we use dev otherwise it will use whatever is configured which should usually be live