Its a mess alright
I've started on the BGS part of the patrol system. Its very messy and will need quite a lot of tidying up. But its a nice little proof of concept. I think we will need to set a max distance maybe as a parameter as there is no point in showing folks that they are 5000ly away from a Canonn controlled system/