New Patrol Configuration
We now have a new patrol spreadsheet that is used to store links to individual patrols which can be in tsv format or json format. The code loops through the patrol list loading he data from the urls.

https://docs.google.com/spreadsheets/d/1rDQInpOdsVd4l3iV9KIVW_wWVKmCfJMt5-iBm7XgVzo/edit?usp=sharing

It is possible to disable and individual patrol with a flag from the patrol list.

Now handles null events.