Using EDSM Structure
Well not quite.  We are converting the edsmbody data from an array to dist using the bodyid as the key. Then scan events get converted to the same format and inserted into the edsm structure. It means that all the codex info can be written just once and it is possible to work with complex relationships betweenbodies