latitude and longitude: use four digit precision
This updates the codex Point of Interest code to use
four digit precision for latitude and longitude.  That
reduces the margin of error for directions by two orders
of magnitude.

Two digit precision was out by up to 100 meters on a
small (two probe) planet; four digit also matches what
the UI in the game displays.