make plugin name consistent in logs and EDMC display
This synchronises the plugin name used in the EDMC UI, as returned from
the `plugin_start{,3}` method, with the name used in the logger.

Since the former was derived from the path to the file, and the later
hard-coded, I opted for the best of both worlds™.  We now use the name
of the directory, but strip the `EDMC-` prefix to keep in sync with
previous behaviour.