fix EDMCOverlay lazy-loading implementation
In some cases the lazy-loading implementation for the EDMCOverlay
plugin did not work, but instead entered a failure loop where it
never *tried* to import the module.

This rewrites the lazy-loading substantially, fixing the problem in the
process and making it overall more robust to errors.  Key changes:

Lazy-loading is now *always* used.  The Canonn plugin will attempt to
start the overlay service the first time a message is displayed, and
not before.

Various internal state tracking is dropped, in favour of the simplest
possible pair: do we have the module, and do we have a connection to it?

The logic around calling `EDMCOverlay.Overlay.connect()` is removed, as
that is now handled internally to the EDMCOverlay plugin.  It is easier
and more robust to let them own the problem.

Error reporting now distinguishes "no EDMCOverlay plugin" from other
errors more robustly, and avoids multiple warnings if the plugin is
not available.

Errors where the EDMCOverlay server connection breaks, such as someone
killing their server process, should now silently restart
our connection.

Any displayed messages will be lost, but new messages should get where
they need to be, which is an improvement over the previous situation:
before this, that would simply break any future message display.