Fix plugin release checks failing if EDMC startup is slow
The plugin release thread was started when EDMC created the Tk widgets
for the UI.  If a new release was found a custom Tk event was queued to
notify the user from the UI thread.

Unfortunately, if EDMC was slow, or github especially efficient, the
release check could finish *before* EDMC had entered the Tk main loop.

That led to an exception from tkinter, since it only waits 1.5 seconds
for the main loop to start before giving up.

This changes the logic to use `.after_idle`, which will start the update
check the next time the Tk main loop is idle.  That ensures the check is
always running *after* the main loop has started.

Users shouldn't experience any significant latency: the main loop will
idle more or less immediately after being entered the first time, and
the user wouldn't see our notification before then in any case.

This fixes this exception:

    Exception in thread Thread-7:
    Traceback (most recent call last):
      File "threading.pyc", line 1038, in _bootstrap_inner
      File "C:\Users\danie\AppData\Local\EDMarketConnector\plugins\E-EDMC-Canonn\canonn\release.py", line 103, in run
        self.release.release_pull()
      File "C:\Users\danie\AppData\Local\EDMarketConnector\plugins\E-EDMC-Canonn\canonn\release.py", line 193, in release_pull
        self.event_generate('<<ReleaseUpdate>>', when='tail')
      File "tkinter\__init__.pyc", line 1914, in event_generate